src/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
faker==37.1.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
mako==1.3.9
markupsafe==3.0.2
passlib==1.7.4
pip==24.3.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.1
typing-extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
uv==0.6.13
uvicorn==0.34.0
werkzeug==3.1.3
